add style to images

+++ app/assets/javascripts/items.coffee
@@ -201,13 +201,14 @@ checkmodeSave = (collection) ->
 ready =  () ->
   #handle image delete links
   $('.photo-delete').click () ->
-    id = $(this).data('photo-id')
-    console.log('deleting photo '+id)
-    $.ajax '/photos/'+id, {
-      method: 'delete',
-      success: ->
-        location.reload()
-    }
+    if confirm("Biztosan törlöd a képet?")
+      id = $(this).data('photo-id')
+      console.log('deleting photo '+id)
+      $.ajax '/photos/'+id, {
+        method: 'delete',
+        success: ->
+          location.reload()
+      }
   #handle backgrid
   gridContainer = $('#grid-container')
   if gridContainer.length > 0
